requirement.txt
torch
torchvision
numpy
opencv-python
pillow
matplotlib
tensorboard
PyYAML
